Move defs outside init into separate file, update build scripts
