+++ scripts/py_matter_idl/files.gni
@@ -22,6 +22,7 @@ matter_idl_generator_templates = [
 matter_idl_generator_sources = [
   "${chip_root}/scripts/py_matter_idl/matter_idl/__init__.py",
   "${chip_root}/scripts/py_matter_idl/matter_idl/backwards_compatibility/__init__.py",
+  "${chip_root}/scripts/py_matter_idl/matter_idl/backwards_compatibility/idl_compatibility.py",
   "${chip_root}/scripts/py_matter_idl/matter_idl/generators/__init__.py",
   "${chip_root}/scripts/py_matter_idl/matter_idl/generators/cpp/__init__.py",
   "${chip_root}/scripts/py_matter_idl/matter_idl/generators/cpp/application/__init__.py",
